+++ src/mouse.h
@@ -4,5 +4,5 @@
 extern void (*mouse_poll)(int x, int y, int b);
 
 extern int mousepos;
-extern int64_t mousedelay;
+extern int mousedelay;
