+++ src/vid_hercules.c
@@ -20,7 +20,7 @@ typedef struct hercules_t
         uint8_t ctrl, ctrl2, stat;
 
         int dispontime, dispofftime;
-        int64_t vidtime;
+        int vidtime;
         
         int firstline, lastline;
