+++ src/vid_incolor.c
@@ -151,7 +151,7 @@ typedef struct incolor_t
         uint8_t ctrl, ctrl2, stat;
 
         int dispontime, dispofftime;
-        int64_t vidtime;
+        int vidtime;
         
         int firstline, lastline;
